# Remove caching from RenameMap