Fixes https://github.com/freechipsproject/firrtl/issues/1933

Currently, `RenameMap` caches every single lookup. This was part of the original Instance Annotations PR (https://github.com/freechipsproject/firrtl/pull/865) because lookups were multi-part and it sped things up a lot. https://github.com/freechipsproject/firrtl/pull/967 removed the recursiveness of the lookup though, so the caching is doing almost nothing but using extra memory these days. I did some (not very thorough) benchmarking

| design | runtime with caching | runtime without caching (this PR) |
| --- | --- | --- |
| small | 42 s | 45 s |
| large | 525 s | 496 s |
| huge | 1309 s | 1207 s |

These are only single runs so this change is probably performance neutral to a very slight win.

### Contributor Checklist

- [x] Did you add Scaladoc to every public function/method?
- [NA] Did you update the FIRRTL spec to include every new feature/behavior?
- [NA] Did you add at least one test demonstrating the PR?
- [x] Did you delete any extraneous printlns/debugging code?
- [x] Did you specify the type of improvement?
- [x] Did you state the API impact?
- [x] Did you specify the code generation impact?
- [x] Did you request a desired merge strategy?
- [x] Did you add text to be included in the Release Notes for this change?

#### Type of Improvement

<!-- Choose one or more from the following: -->
  - bug fix
  - performance improvement (ie. large designs use a *lot* less memory)


#### API Impact

No impact

#### Backend Code Generation Impact

No impact

#### Desired Merge Strategy

 - Squash

#### Release Notes
Fix memory use bug in LowerTypes  (#1933)

### Reviewer Checklist (only modified by reviewer)
- [ ] Did you add the appropriate labels?
- [ ] Did you mark the proper milestone (1.2.x, 1.3.0, 1.4.0) ?
- [ ] Did you review?
- [ ] Did you check whether all relevant Contributor checkboxes have been checked?
- [ ] Did you mark as `Please Merge`?